# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "API documentation for the School Management System"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "paths": {
...
    "/attendance/{id}": {
      "get": {
        "summary": "Get an attendance record by ID",
        "description": "Retrieve a single attendance record by its ID",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with attendance record details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Attendance"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "summary": "Update an attendance record",
        "description": "Update an attendance record's information by its ID",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Attendance record updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Attendance"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete an attendance record",
        "description": "Delete an attendance record by its ID",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance record deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Attendance record deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/attendance": {
      "get": {
        "summary": "Get all attendance records",
        "description": "Retrieve a list of all attendance records with optional filtering, sorting, pagination, and search capabilities",
        "tags": [
          "Attendance"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/fieldsParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "in": "query",
            "name": "studentId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by student ID"
          },
          {
            "in": "query",
            "name": "courseId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by course ID"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "present",
                "absent",
                "late",
                "excused"
              ]
            },
            "description": "Filter by attendance status"
          },
          {
            "in": "query",
            "name": "date",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by specific date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "date[gte]",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by date greater than or equal to (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "date[lte]",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by date less than or equal to (YYYY-MM-DD)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with list of attendance records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 10
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Attendance"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new attendance record",
        "description": "Create a new attendance record with the provided data",
        "tags": [
          "Attendance"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttendanceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attendance record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Attendance record created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Attendance"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Attendance": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c89"
          },
          "studentId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "UserId": {
                "type": "string"
              }
            }
          },
          "courseId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "courseCode": {
                "type": "string"
              },
              "title": {
                "type": "string"
              }
            }
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-04-15"
          },
          "status": {
            "type": "string",
            "enum": [
              "present",
              "absent",
              "late",
              "excused"
            ],
            "example": "present"
          },
          "notes": {
            "type": "string",
            "example": "Student arrived 5 minutes late"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "studentId",
          "courseId",
          "date",
          "status"
        ]
      },
      "AttendanceInput": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c90"
          },
          "courseId": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c91"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-04-15"
          },
          "status": {
            "type": "string",
            "enum": [
              "present",
              "absent",
              "late",
              "excused"
            ],
            "example": "present"
          },
          "notes": {
            "type": "string",
            "example": "Student arrived 5 minutes late"
          }
        },
        "required": [
          "studentId",
          "courseId",
          "date",
          "status"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100,
            "description": "Total number of records"
          },
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "example": 25,
            "description": "Number of records per page"
          },
          "pages": {
            "type": "integer",
            "example": 4,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        }
      },
...
    },
    "parameters": {
      "sortParam": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string"
        },
        "description": "Sort order (prefix field with - for descending) e.g. -createdAt,name"
      },
      "fieldsParam": {
        "in": "query",
        "name": "fields",
        "schema": {
          "type": "string"
        },
        "description": "Fields to include (comma separated) e.g. name,code,description"
      },
      "pageParam": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "limitParam": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "default": 25,
          "maximum": 100
        },
        "description": "Number of records per page (max 100)"
      },
...
      "idParam": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Resource ID"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - validation failed",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Validation failed"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Name is required",
                    "Invalid status value"
                  ]
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}